Add files via upload
I realized I had to push the merge on github.desktop. I'm not entirely sure why, but it wasn't letting me merge. It kept saying I had an error with Lab 4 and I had to force the merge another way as it wasn't allowing me to automatically merge the changes/commits. But regardless, figured it out, apologies that I couldn't figure this out during lecture...